Fix intermittent test failures that are based on timing queries.
This fix changes the `queue.get()` functionality slightly by using
'less than or equal to' when comparing to the current time.
This ensures that any newly added jobs to a queue can be retrieved
straight away i.e. a test can add a job, then get it without having to
wait until a small amount of time has passed and the job is 'visible'