Fail gracefully when adding an empty array
Currently when adding an empty array using `queue.add` an error will
be thrown by MongoDB (at least in version 2.x.x), this performs a
check on the payload of `add`. If it's an array with a length of 0
an error will be returned in the callback instead of the Mongo error
being thrown.