✨ Add a ttl option for auto expiry of deleted messages
This change allows a ttl to be set for automatically removing deleted
queue messages. This can be useful for limiting the amount of space used
by queues, while still maintaining some level of auditing.

A new `ttl` option can be specified, in seconds, for how long before
expiring a deleted message. The default is to not have any expiration
i.e. the existing behaviour.

Some extra assertions are added to the existing indexes test.
A new test for the ttl index is also added.